Rename pathview_server to pathview and add convert API
