+++ package.json
@@ -19,7 +19,7 @@
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "eslint .",
     "format": "prettier --write .",
-    "server": "python -m pathview_server.app",
+    "server": "python -m pathview.app",
     "build:package": "python scripts/build_package.py"
   },
   "devDependencies": {
