+++ pathview/__main__.py
@@ -1,0 +1,6 @@
+"""Entry point for: python -m pathview"""
+
+from pathview.cli import main
+
+if __name__ == "__main__":
+    main()
