--- pathview_server/__main__.py
@@ -1,6 +1,0 @@
-"""Entry point for: python -m pathview_server"""
-
-from pathview_server.cli import main
-
-if __name__ == "__main__":
-    main()
